Cleaner feature builds and tests, move test_support to distinct feature
